+++ docs/source/others/changelog.rst
@@ -3,8 +3,14 @@ Changelog
 ==========
 
 
+
 2024.02.3+dev  (XXXX-XX-XX)
 ---------------------------
+**Improvements:**
+
+- Add "In progress" status to source
+- Source list endpoint now also return ids of related layer
+- Add author to Source
 
 **Documentation:**